Move simsys -> sys/
