+++ Makefile
@@ -47,7 +47,7 @@ else
           LIBS += -lmingw32
         endif
       endif
-      SOURCES += simsys_w32_png.cc
+      SOURCES += sys/simsys_w32_png.cc
       CFLAGS  += -DNOMINMAX -DWIN32_LEAN_AND_MEAN -DWINVER=0x0501 -D_WIN32_IE=0x0500
       LIBS    += -lgdi32 -lwinmm -lws2_32 -limm32
       # Disable the console on Windows unless WIN32_CONSOLE is set or graphics are disabled
@@ -267,6 +267,7 @@ SOURCES += display/font.cc
 SOURCES += display/simgraph$(COLOUR_DEPTH).cc
 SOURCES += display/simview.cc
 SOURCES += display/viewport.cc
+SOURCES += finder/placefinder.cc
 SOURCES += freight_list_sorter.cc
 SOURCES += gui/ai_option_t.cc
 SOURCES += gui/banner.cc
@@ -464,12 +465,11 @@ SOURCES += simplan.cc
 SOURCES += simsignalbox.cc
 SOURCES += simskin.cc
 SOURCES += simsound.cc
-SOURCES += simsys.cc
 SOURCES += simticker.cc
 SOURCES += simtool.cc
 SOURCES += simware.cc
 SOURCES += simworld.cc
-SOURCES += finder/placefinder.cc
+SOURCES += sys/simsys.cc
 SOURCES += unicode.cc
 SOURCES += utils/cbuffer_t.cc
 SOURCES += utils/csv.cc
@@ -498,7 +498,7 @@ SOURCES += dataobj/livery_scheme.cc
 SOURCES += dataobj/replace_data.cc
 
 ifeq ($(BACKEND),allegro)
-  SOURCES  += simsys_d.cc
+  SOURCES  += sys/simsys_d.cc
   SOURCES += sound/allegro_sound.cc
   SOURCES += music/allegro_midi.cc
   ifeq ($(ALLEGRO_CONFIG),)
@@ -513,14 +513,14 @@ ifeq ($(BACKEND),allegro)
 endif
 
 ifeq ($(BACKEND),gdi)
-  SOURCES += simsys_w.cc
+  SOURCES += sys/simsys_w.cc
   SOURCES += music/w32_midi.cc
   SOURCES += sound/win32_sound.cc
   CFLAGS += -DGDI_SOUND
 endif
 
 ifeq ($(BACKEND),sdl)
-  SOURCES += simsys_s.cc
+  SOURCES += sys/simsys_s.cc
   ifeq ($(OSTYPE),mac)
 		ifeq ($(shell expr $(AV_FOUNDATION) \>= 1), 1)
 			# Core Audio (AVFoundation) base sound system routines
@@ -562,7 +562,7 @@ ifeq ($(BACKEND),sdl)
 endif
 
 ifeq ($(BACKEND),sdl2)
-  SOURCES += simsys_s2.cc
+  SOURCES += sys/simsys_s2.cc
   ifeq ($(OSTYPE),mac)
 		ifeq ($(shell expr $(AV_FOUNDATION) \>= 1), 1)
 			# Core Audio (AVFoundation) base sound system routines
@@ -600,7 +600,7 @@ ifeq ($(BACKEND),sdl2)
 endif
 
 ifeq ($(BACKEND),mixer_sdl2)
-  SOURCES += simsys_s2.cc
+  SOURCES += sys/simsys_s2.cc
   ifeq ($(SDL2_CONFIG),)
     ifeq ($(OSTYPE),mac)
       SDL_CFLAGS  := -I/Library/Frameworks/SDL2.framework/Headers
@@ -628,7 +628,7 @@ ifeq ($(BACKEND),mixer_sdl2)
 endif
 
 ifeq ($(BACKEND),mixer_sdl)
-  SOURCES += simsys_s.cc
+  SOURCES += sys/simsys_s.cc
   SOURCES += sound/sdl_mixer_sound.cc
   SOURCES += music/sdl_midi.cc
   ifeq ($(SDL_CONFIG),)
@@ -648,7 +648,7 @@ ifeq ($(BACKEND),mixer_sdl)
 endif
 
 ifeq ($(BACKEND),opengl)
-  SOURCES += simsys_opengl.cc
+  SOURCES += sys/simsys_opengl.cc
   ifeq ($(OSTYPE),mac)
 		ifeq ($(shell expr $(AV_FOUNDATION) \>= 1), 1)
 			# Core Audio (AVFoundation) base sound system routines
@@ -694,7 +694,7 @@ ifeq ($(BACKEND),opengl)
 endif
 
 ifeq ($(BACKEND),posix)
-  SOURCES += simsys_posix.cc
+  SOURCES += sys/simsys_posix.cc
   SOURCES += music/no_midi.cc
   SOURCES += sound/no_sound.cc
 endif
