+++ clipboard_internal.cc
@@ -5,7 +5,7 @@
 
 #include <string.h>
 
-#include "simsys.h"
+#include "sys/simsys.h"
 #include "display/simgraph.h"
 #include "simdebug.h"
