+++ dataobj/route.cc
@@ -41,7 +41,7 @@
 
 
 #ifdef DEBUG_ROUTES
-#include "../simsys.h"
+#include "../sys/simsys.h"
 #endif
 
 bool route_t::suspend_private_car_routing = false;
@@ -587,7 +587,7 @@ bool route_t::find_route(karte_t *welt, const koord3d start, test_driver_t *tdri
 						}
 						else if(tmp->parent->dir!=tmp->dir  &&  tmp->parent->parent!=NULL)
 						{
-							// discourage 90° turns
+							// discourage 90ï¿½ turns
 							k->g += 10;
 						}
 					}
@@ -956,7 +956,7 @@ route_t::route_result_t route_t::intern_calc_route(karte_t *welt, const koord3d 
 					if (tmp->dir != current_dir) {
 						new_g += 30;
 						if (tmp->parent->dir != tmp->dir  &&  tmp->parent->parent != NULL) {
-							// discourage 90° turns
+							// discourage 90ï¿½ turns
 							new_g += 10;
 						}
 						else if (ribi_t::is_perpendicular(tmp->dir, current_dir))
@@ -1199,7 +1199,7 @@ void route_t::postprocess_water_route(karte_t *welt)
 /* searches route, uses intern_calc_route() for distance between stations
  * handles only driving in stations by itself
  * corrected 12/2005 for station search
- * @author Hansjörg Malthaner, prissi
+ * @author Hansjï¿½rg Malthaner, prissi
  */
  route_t::route_result_t route_t::calc_route(karte_t *welt, const koord3d start, const koord3d ziel, test_driver_t* const tdriver, const sint32 max_khm, const uint32 axle_load, bool is_tall, sint32 max_len, const sint64 max_cost, const uint32 convoy_weight, koord3d avoid_tile, uint8 direction, find_route_flags flags)
 {
