+++ display/font.cc
@@ -236,7 +236,7 @@ static bool dsp_read_bdf_font(FILE* fin, font_type* font)
 
 #ifdef USE_FREETYPE
 #include "../gui/gui_theme.h"
-#include "../simsys.h"
+#include "../sys/simsys.h"
 
 #include <ft2build.h>
 #include FT_FREETYPE_H
