+++ gui/pakselector.cc
@@ -8,7 +8,7 @@
 #include "pakselector.h"
 #include "../dataobj/environment.h"
 #include "../utils/cbuffer_t.h"
-#include "../simsys.h"
+#include "../sys/simsys.h"
