+++ gui/sprachen.cc
@@ -21,7 +21,7 @@
 
 #include "../dataobj/environment.h"
 #include "../dataobj/translator.h"
-#include "../simsys.h"
+#include "../sys/simsys.h"
 #include "../utils/simstring.h"
 
 #define L_DIALOG_WIDTH (220)
