+++ sys/simsys.cc
@@ -19,8 +19,8 @@
 #define NO_UINT64_TYPES
 #endif
 
-#include "macros.h"
-#include "simmain.h"
+#include "../macros.h"
+#include "../simmain.h"
 #include "simsys.h"
